+++ waspc/data/Generator/templates/sdk/wasp/server/email/core/types.ts
@@ -1,6 +1,6 @@
 {{={= =}=}}
 // PRIVATE API
-export type EmailProvider = SMTPEmailProvider | SendGridProvider | MailgunEmailProvider | ResendEmailProvider | DummyEmailProvider;
+export type EmailProvider = SMTPEmailProvider | SendGridProvider | MailgunEmailProvider | ResendEmailProvider | MailtrapEmailProvider | DummyEmailProvider;
 
 // PRIVATE API
 export type SMTPEmailProvider = {
@@ -29,6 +29,14 @@ export type MailgunEmailProvider = {
 export type ResendEmailProvider = {
   type: "resend";
   apiKey: string;
+};
+
+// PRIVATE API
+export type MailtrapEmailProvider = {
+  type: "mailtrap";
+  apiToken: string;
+  sandbox?: boolean;
+  testInboxId?: number;
 };
 
 // PRIVATE API