+++ waspc/data/packages/spec/src/spec/publicApi/waspSpec.ts
@@ -747,6 +747,7 @@ export type EmailSenderProviderName =
   | "SendGrid"
   | "Mailgun"
   | "Resend"
+  | "Mailtrap"
   | "Dummy";
 
 /**