+++ waspc/src/Wasp/AppSpec/App/EmailSender.hs
@@ -19,7 +19,7 @@ data EmailSender = EmailSender
   }
   deriving (Show, Eq, Data, Generic, FromJSON)
 
-data EmailProvider = SMTP | SendGrid | Mailgun | Resend | Dummy
+data EmailProvider = SMTP | SendGrid | Mailgun | Resend | Mailtrap | Dummy
   deriving (Show, Eq, Data, Generic, FromJSON)
 
 data EmailFromField = EmailFromField